requirements.txt
docutils==0.20.1
intel_tensorflow==2.9.1
numpy==1.23.5
Pillow==9.4.0
streamlit==1.22.0
